Converted all accuracy values to floats in evaluation script

+++ src/evaluation.py
@@ -171,9 +171,9 @@ class Evaluator(object):
         log_lat_dis = []
         if params.n_lat_dis:
             lat_dis_accu = self.eval_lat_dis_accuracy()
-            log_lat_dis.append(('lat_dis_accu', np.mean(lat_dis_accu)))
+            log_lat_dis.append(('lat_dis_accu', float(np.mean(lat_dis_accu))))
             for accu, (name, _) in zip(lat_dis_accu, params.attr):
-                log_lat_dis.append(('lat_dis_accu_%s' % name, accu))
+                log_lat_dis.append(('lat_dis_accu_%s' % name, float(accu)))
             logger.info('Latent discriminator accuracy:')
             print_accuracies(log_lat_dis)
 
@@ -183,10 +183,10 @@ class Evaluator(object):
             ptc_dis_real_preds, ptc_dis_fake_preds = self.eval_ptc_dis_accuracy()
             accu_real = (np.array(ptc_dis_real_preds).astype(np.float32) >= 0.5).mean()
             accu_fake = (np.array(ptc_dis_fake_preds).astype(np.float32) <= 0.5).mean()
-            log_ptc_dis.append(('ptc_dis_preds_real', np.mean(ptc_dis_real_preds)))
-            log_ptc_dis.append(('ptc_dis_preds_fake', np.mean(ptc_dis_fake_preds)))
-            log_ptc_dis.append(('ptc_dis_accu_real', accu_real))
-            log_ptc_dis.append(('ptc_dis_accu_fake', accu_fake))
+            log_ptc_dis.append(('ptc_dis_preds_real', float(np.mean(ptc_dis_real_preds))))
+            log_ptc_dis.append(('ptc_dis_preds_fake', float(np.mean(ptc_dis_fake_preds))))
+            log_ptc_dis.append(('ptc_dis_accu_real', float(accu_real)))
+            log_ptc_dis.append(('ptc_dis_accu_fake', float(accu_fake)))
             log_ptc_dis.append(('ptc_dis_accu', (accu_real + accu_fake) / 2))
             logger.info('Patch discriminator accuracy:')
             print_accuracies(log_ptc_dis)
@@ -198,8 +198,8 @@ class Evaluator(object):
             k = 0
             log_clf_dis += [('clf_dis_accu', np.mean(clf_dis_accu))]
             for name, n_cat in params.attr:
-                log_clf_dis.append(('clf_dis_accu_%s' % name, np.mean(clf_dis_accu[k:k + n_cat])))
-                log_clf_dis.extend([('clf_dis_accu_%s_%i' % (name, j), clf_dis_accu[k + j])
+                log_clf_dis.append(('clf_dis_accu_%s' % name, float(np.mean(clf_dis_accu[k:k + n_cat]))))
+                log_clf_dis.extend([('clf_dis_accu_%s_%i' % (name, j), float(clf_dis_accu[k + j]))
                                     for j in range(n_cat)])
                 k += n_cat
             logger.info('Classifier discriminator accuracy:')
@@ -211,8 +211,8 @@ class Evaluator(object):
         k = 0
         log_clf += [('clf_accu', np.mean(clf_accu))]
         for name, n_cat in params.attr:
-            log_clf.append(('clf_accu_%s' % name, np.mean(clf_accu[k:k + n_cat])))
-            log_clf.extend([('clf_accu_%s_%i' % (name, j), clf_accu[k + j])
+            log_clf.append(('clf_accu_%s' % name, float(np.mean(clf_accu[k:k + n_cat]))))
+            log_clf.extend([('clf_accu_%s_%i' % (name, j), float(clf_accu[k + j]))
                             for j in range(n_cat)])
             k += n_cat
         logger.info('Classifier accuracy:')
@@ -223,8 +223,8 @@ class Evaluator(object):
 
         # JSON log
         to_log = dict([
-            ('n_epoch', n_epoch),
-            ('ae_loss', ae_loss)
+            ('n_epoch', float(n_epoch)),
+            ('ae_loss', float(ae_loss))
         ] + log_lat_dis + log_ptc_dis + log_clf_dis + log_clf)
         logger.debug("__log__:%s" % json.dumps(to_log))
